+++ mayan/apps/cabinets/events.py
@@ -6,9 +6,16 @@ from mayan.apps.events.classes import EventTypeNamespace
 
 namespace = EventTypeNamespace(label=_('Cabinets'), name='cabinets')
 
-event_cabinets_add_document = namespace.add_event_type(
+
+event_cabinet_created = namespace.add_event_type(
+    label=_('Cabinet created'), name='cabinet_created'
+)
+event_cabinet_edited = namespace.add_event_type(
+    label=_('Cabinet edited'), name='cabinet_edited'
+)
+event_cabinet_add_document = namespace.add_event_type(
     label=_('Document added to cabinet'), name='add_document'
 )
-event_cabinets_remove_document = namespace.add_event_type(
+event_cabinet_remove_document = namespace.add_event_type(
     label=_('Document removed from cabinet'), name='remove_document'
 )